refactor: Extract a lib
